+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/elle/turing/mod-3/project/purrfect-status/src/index.js":"1","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/App/index.js":"2","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Home/index.js":"3","/Users/elle/turing/mod-3/project/purrfect-status/src/apiCalls.js":"4","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Form/index.js":"5","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Picture/index.js":"6"},{"size":276,"mtime":1610736339811,"results":"7","hashOfConfig":"8"},{"size":166,"mtime":1610736339806,"results":"9","hashOfConfig":"8"},{"size":669,"mtime":1610737390683,"results":"10","hashOfConfig":"8"},{"size":336,"mtime":1610736339811,"results":"11","hashOfConfig":"8"},{"size":1466,"mtime":1610736339808,"results":"12","hashOfConfig":"8"},{"size":1429,"mtime":1610737425755,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1swfd7p",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/elle/turing/mod-3/project/purrfect-status/src/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/App/index.js",[],["27","28"],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Home/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/apiCalls.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Form/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Picture/index.js",["29"],{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","severity":1,"message":"35","line":17,"column":47,"nodeType":"36","endLine":17,"endColumn":49,"suggestions":"37"},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'generateRandomCatPicture'. Either include it or remove the dependency array.","ArrayExpression",["40"],"no-global-assign","no-unsafe-negation",{"desc":"41","fix":"42"},"Update the dependencies array to be: [generateRandomCatPicture]",{"range":"43","text":"44"},[525,527],"[generateRandomCatPicture]"]
+[{"/Users/elle/turing/mod-3/project/purrfect-status/src/index.js":"1","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/App/index.js":"2","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Home/index.js":"3","/Users/elle/turing/mod-3/project/purrfect-status/src/apiCalls.js":"4","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Form/index.js":"5","/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Picture/index.js":"6"},{"size":276,"mtime":1610736339811,"results":"7","hashOfConfig":"8"},{"size":754,"mtime":1610738149630,"results":"9","hashOfConfig":"8"},{"size":389,"mtime":1610738203382,"results":"10","hashOfConfig":"8"},{"size":336,"mtime":1610736339811,"results":"11","hashOfConfig":"8"},{"size":1466,"mtime":1610736339808,"results":"12","hashOfConfig":"8"},{"size":1429,"mtime":1610737589386,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1swfd7p",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"27","usedDeprecatedRules":"22"},"/Users/elle/turing/mod-3/project/purrfect-status/src/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/App/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Home/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/apiCalls.js",[],["28","29"],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Form/index.js",[],"/Users/elle/turing/mod-3/project/purrfect-status/src/Components/Picture/index.js",["30"],"import React, { useState, useEffect } from 'react';\nimport { getCatPicture } from '../../apiCalls';\nimport './Picture.scss'\nimport { IoPawSharp } from 'react-icons/io5';\n\nconst Picture = ({ setUserInput, userInput }) => {\n  const [pictureUrl, setPictureUrl] = useState('');\n  const [isFavorite, setIsFavorite] = useState(false);\n\n  const generateRandomCatPicture = () => {\n    getCatPicture().then(data =>{\n      setPictureUrl(data[0].url);\n      setUserInput(null);\n    })\n  }\n\n  useEffect(() => generateRandomCatPicture(), [])\n\n  return (\n    <section className=\"picture-section\">\n      <div \n        className=\"cat-picture-wrapper\" \n        style={{\n          backgroundImage: `url(${pictureUrl})`,\n          backgroundPosition: \"center\",\n          backgroundSize: \"cover\",\n          backgroundRepeat: \"no-repeat\"\n        }}>\n        <div className=\"icon-wrapper\">\n          {userInput && <IoPawSharp \n            className=\"icon\" \n            onClick={() => setIsFavorite(prevStatus => !prevStatus)} \n            style={{color: isFavorite ? '#177fcc' : 'white' }}\n          />}\n        </div>\n        <figure className=\"dispay-inputs-wrapper\">\n          <p>{userInput && userInput.statusCode}</p>\n          <p>{userInput && userInput.explaination}</p>\n        </figure>\n      </div>\n      <div>\n        <button onClick={generateRandomCatPicture}>Not Purrfect</button>\n      </div>\n    </section>\n  )\n}\n\nexport default Picture",{"ruleId":"31","replacedBy":"32"},{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","severity":1,"message":"36","line":17,"column":47,"nodeType":"37","endLine":17,"endColumn":49,"suggestions":"38"},"no-native-reassign",["39"],"no-negated-in-lhs",["40"],"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'generateRandomCatPicture'. Either include it or remove the dependency array.","ArrayExpression",["41"],"no-global-assign","no-unsafe-negation",{"desc":"42","fix":"43"},"Update the dependencies array to be: [generateRandomCatPicture]",{"range":"44","text":"45"},[525,527],"[generateRandomCatPicture]"]
